+++ tests/webhook.php
@@ -36,7 +36,7 @@ class Test
         }
     }
 
-    private function example()
+    private static function example()
     {
         return new Webhook([
             "url" => "https://webhook.site/60e9c18e-4b5c-4369-bda1-ab5fcd8e1b29",